refactor: completion to dedicated mod file
